Fix dependency violation: move getGitRepository to common module
- Move getGitRepository function from src/rmfilter/utils.ts to src/common/git.ts
- Update imports in src/common/github/identifiers.ts and pull_requests.ts 
- Remove stale setDebug imports from test files
- Resolves improper dependency where common modules imported from feature modules
