+++ src/common/github/pull_requests.ts
@@ -5,7 +5,7 @@ import type { DetailedReviewComment } from '../../rmpr/types.ts';
 import { debugLog, error, log, warn } from '../../logging.ts';
 import { parsePrOrIssueNumber } from './identifiers.ts';
 import { getCurrentBranchName } from '../git.ts';
-import { getGitRepository } from '../../rmfilter/utils.ts';
+import { getGitRepository } from '../git.js';
 
 export interface CommentAuthor {
   login: string;
